Partie 1 fonctionnelle

+++ Backend/controllers/BookController.js
@@ -1,5 +1,8 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
+
+// Créer un livre
 exports.createBook = (req, res, next) => {
   console.log('Request body:', req.body);
   console.log('Request file:', req.file);
@@ -25,6 +28,7 @@ exports.createBook = (req, res, next) => {
   }
 };
 
+// Modifier un livre
 exports.modifyBook = (req, res, next) => {
   const bookObject = req.file ? {
       ...JSON.parse(req.body.book),
@@ -35,7 +39,7 @@ exports.modifyBook = (req, res, next) => {
   Book.findOne({_id: req.params.id})
       .then((book) => {
           if (book.userId != req.auth.userId) {
-              res.status(401).json({ message : 'Not authorized'});
+              res.status(401).json({ message : 'Non-autorisé'});
           } else {
               Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
               .then(() => res.status(200).json({message : 'Objet modifié!'}))
@@ -46,20 +50,39 @@ exports.modifyBook = (req, res, next) => {
       });
 };
 
+// Supprimer un livre
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-      .then(() => res.status(200).json({ message: 'Livre supprimé !'}))
-      .catch(error => res.status(400).json({ error })
-      );
+  Book.findOne({ _id: req.params.id})
+      .then(book => {
+          if (book.userId != req.auth.userId) {
+              res.status(401).json({message: 'Non-autorisé'});
+          } else {
+              const filename = book.imageUrl.split('/images/')[1];
+              fs.unlink(`images/${filename}`, () => {
+                  Book.deleteOne({_id: req.params.id})
+                      .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                      .catch(error => res.status(401).json({ error }));
+              });
+          }
+      })
+      .catch( error => {
+          res.status(500).json({ error });
+      });
 };
 
+// Récupérer un livre
 exports.getOneBook = (req, res, next) => {
     Book.findOne({ _id: req.params.id })
-      .then(book => res.status(200).json(book))
-      .catch(error => res.status(404).json({ error })
-      );
+      .then(book => {
+        if (!book) {
+          return res.status(404).json({ message: 'Livre non trouvé' });
+        }
+        res.status(200).json(book);
+      })
+      .catch(error => res.status(500).json({ error: 'Erreur serveur' }));
 };
 
+// Récupérer tous les livres
 exports.getAllBooks = (req, res, next) => {
     Book.find()
       .then(books => res.status(200).json(books))
@@ -67,6 +90,7 @@ exports.getAllBooks = (req, res, next) => {
       );
 };
 
+// Récupérer les meilleurs livres
 exports.getBestRatedBooks = (req, res, next) => {
   Book.find().sort({ averageRating: -1 }).limit(5)
     .then(books => res.status(200).json(books))
